# Simplify a SQL query for getting the best_trial's trial_id